% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_network.R
\name{merge_network}
\alias{merge_network}
\title{Merge Reactome, KEGG, and HMDB network and consolidate unique pairs}
\usage{
merge_network(
  reactome.edge,
  reactome.node,
  kegg.edge,
  kegg.node,
  hmdb.edge,
  hmdb.node,
  write_output = FALSE,
  output_dir = "outputs"
)
}
\arguments{
\item{reactome.edge}{Reactome edge attribute file}

\item{reactome.node}{Reactome node attribute file}

\item{kegg.edge}{Reactome edge attribute file}

\item{kegg.node}{Reactome node attribute file}

\item{hmdb.edge}{Reactome edge attribute file}

\item{hmdb.node}{Reactome node attribute file}

\item{write_output}{Should output files be written? (default: FALSE)}

\item{output_dir}{Output directory path (default: "outputs")}
}
\value{
List containing node attributes, edge attributes, and solid edges
}
\description{
Merge Reactome, KEGG, and HMDB network and consolidate unique pairs
}
