% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_input.R
\name{standardize_input}
\alias{standardize_input}
\title{Standardize column names using the provided meta information file}
\usage{
standardize_input(data, info)
}
\arguments{
\item{data}{Input dataset containing merged metabolic and transcriptomic data}

\item{info}{Metadata for column standardization}
}
\description{
Standardize column names using the provided meta information file
}
